requirements.txt
discord.py==2.1.0
aiohttp==3.8.1
python-dotenv==1.0.0
langdetect==1.0.9
loguru==0.6.0